refactor: 컨트롤러 3개로 분리
