+++ src/js/index.js
@@ -1,42 +1,40 @@
 import {
-  $modalClose,
   $modal,
-  $lottoNumbersToggleButton,
+  $modalClose,
   $priceSubmitForm,
+  $lottoNumbersToggleButton,
   $winningNumberForm,
   $restartButton,
   $manualPurchaseButton,
   $manualPurchaseForm,
-  $autoPurchaseButton,
   $manualPurchaseDetail,
+  $autoPurchaseButton,
 } from "./elements.js";
 import { onModalClose, showElement } from "./utils.js";
+import LottoModel from "./lotto/LottoModel.js";
+import LottoView from "./lotto/LottoView.js";
 import LottoController from "./lotto/LottoController.js";
+import LottoPurchaseController from "./lotto/LottoPurchaseController.js";
+import LottoConfirmationController from "./lotto/LottoConfirmationController.js";
 
-const lottoController = new LottoController();
+const lottoModel = new LottoModel();
+const lottoView = new LottoView();
 
+const lottoController = new LottoController(lottoModel, lottoView);
+const lottoPurchaseController = new LottoPurchaseController(
+  lottoModel,
+  lottoView
+);
+const lottoConfirmationController = new LottoConfirmationController(
+  lottoModel,
+  lottoView
+);
+
+// lottoController관련 handlers
 $priceSubmitForm.addEventListener("submit", (e) => {
   e.preventDefault();
 
   lottoController.onSubmitPrice(e.target.elements["price-input"].value);
-});
-
-$manualPurchaseButton.addEventListener("click", () => {
-  showElement($manualPurchaseDetail);
-});
-
-$manualPurchaseForm.addEventListener("submit", (e) => {
-  e.preventDefault();
-
-  lottoController.onSubmitManualPurchaseNumber(
-    Array.from(e.target.elements["manual-purchase-number"]).map((v) =>
-      Number(v.value)
-    )
-  );
-});
-
-$autoPurchaseButton.addEventListener("click", () => {
-  lottoController.onClickAutoPurchaseButton();
 });
 
 $lottoNumbersToggleButton.addEventListener(
@@ -44,17 +42,40 @@ $lottoNumbersToggleButton.addEventListener(
   lottoController.onToggleLottoNumbers.bind(lottoController)
 );
 
+// lottoPurchaseController관련 handlers
+$manualPurchaseButton.addEventListener("click", () => {
+  showElement($manualPurchaseDetail);
+});
+
+$manualPurchaseForm.addEventListener("submit", (e) => {
+  e.preventDefault();
+
+  lottoPurchaseController.onSubmitManualPurchaseNumber(
+    Array.from(e.target.elements["manual-purchase-number"]).map((v) =>
+      Number(v.value)
+    )
+  );
+});
+
+$autoPurchaseButton.addEventListener("click", () => {
+  lottoPurchaseController.onClickAutoPurchaseButton();
+});
+
+// confirmationController관련 handlers
 $winningNumberForm.addEventListener("submit", (e) => {
   e.preventDefault();
 
-  lottoController.onSubmitResultNumber(
-    Array.from(e.target.elements["winning-number"]).map((v) => Number(v.value)),
-    Number(e.target.elements["bonus-number"].value)
-  );
+  const winningNumber = Array.from(
+    e.target.elements["winning-number"]
+  ).map((v) => Number(v.value));
+  const bonusNumber = Number(e.target.elements["bonus-number"].value);
+
+  lottoController.onSubmitResultNumber(winningNumber, bonusNumber);
+  lottoConfirmationController.onSubmitResultNumber(winningNumber, bonusNumber);
 });
 
 $modalClose.addEventListener("click", () => onModalClose($modal));
 
 $restartButton.addEventListener("click", () => {
-  lottoController.onClickRestartButton();
+  lottoConfirmationController.onClickRestartButton();
 });
