+++ src/js/lotto/LottoController.js
@@ -1,124 +1,26 @@
-import LottoModel from "./LottoModel.js";
-import LottoView from "./LottoView.js";
-import {
-  hideElement,
-  onModalShow,
-  onModalClose,
-  resetInput,
-} from "../utils.js";
-import { $modal, $manualPurchaseDetail, $priceInput } from "../elements.js";
+import { onModalShow } from "../utils.js";
+import { $modal } from "../elements.js";
 import {
   INVALID_PRICE_ERROR,
   INVALID_WINNGNUMBER_ERROR,
   DUPLICATED_WINNINGNUMBER_ERROR,
 } from "../constants.js";
-
+import {
+  isValidPrice,
+  isNumbersInRange,
+  isDistinctNumbers,
+} from "../validates.js";
 export default class LottoController {
-  constructor() {
-    this.lottoModel = new LottoModel();
-    this.lottoView = new LottoView();
-    this.prizeTable = {
-      ranking1: {
-        num: 0,
-        prize: 2000000000,
-        condition: "6개",
-      },
-      ranking2: {
-        num: 0,
-        prize: 30000000,
-        condition: "5개 + 보너스볼",
-      },
-      ranking3: {
-        num: 0,
-        prize: 1500000,
-        condition: "5개",
-      },
-      ranking4: {
-        num: 0,
-        prize: 50000,
-        condition: "4개",
-      },
-      ranking5: {
-        num: 0,
-        prize: 5000,
-        condition: "3개",
-      },
-      noPrize: {
-        num: 0,
-        prize: 0,
-        condition: "2개 이하",
-      },
-    };
-  }
-
-  isValidPrice(price) {
-    return price > 0 && price % 1000 === 0; // price는 1000원 단위의 양수여야 한다.
-  }
-
-  isNumbersInRange(numbers, min, max) {
-    return numbers.every((num) => min <= num && max >= num);
-  }
-
-  isDistinctNumbers(numbers) {
-    const numbersSet = new Set(numbers);
-
-    return numbersSet.size === numbers.length;
-  }
-
-  countMatchedNumbers(lottoNumber, resultNumber) {
-    const matchedNumbers = lottoNumber.filter(
-      (num) => resultNumber.indexOf(num) !== -1
-    );
-
-    return matchedNumbers.length;
-  }
-
-  checkRanking(lottoNumber, winningNumber, bonusNumber) {
-    const numOfMatched = this.countMatchedNumbers(lottoNumber, winningNumber);
-    switch (numOfMatched) {
-      case 3:
-        return "ranking5";
-      case 4:
-        return "ranking4";
-      case 5:
-        if (this.countMatchedNumbers(lottoNumber, [bonusNumber]) === 1) {
-          return "ranking2";
-        }
-        return "ranking3";
-      case 6:
-        return "ranking1";
-      default:
-        return "noPrize";
-    }
-  }
-
-  setPrizeTable(winningNumber, bonusNumber) {
-    this.lottoModel.lottoList.forEach((lotto) => {
-      const ranking = this.checkRanking(
-        lotto.number,
-        winningNumber,
-        bonusNumber
-      );
-      this.prizeTable[ranking].num++;
-    });
-  }
-
-  calculateEarningRate() {
-    const totalPrize = Object.values(this.prizeTable).reduce(
-      (totalPrize, ranking) => {
-        return (totalPrize += ranking.num * ranking.prize);
-      },
-      0
-    );
-
-    return Math.round((totalPrize / this.lottoModel.price) * 100);
+  constructor(lottoModel, lottoView) {
+    this.lottoModel = lottoModel;
+    this.lottoView = lottoView;
   }
 
   onSubmitPrice(price) {
     this.lottoView.resetLottoView(); // 구입 금액 재입력 했을 경우
     this.lottoModel.resetLottoList();
 
-    if (!this.isValidPrice(price)) {
+    if (!isValidPrice(price)) {
       alert(INVALID_PRICE_ERROR);
       this.lottoView.resetLottoView();
 
@@ -139,59 +41,17 @@ export default class LottoController {
 
   onSubmitResultNumber(winningNumber, bonusNumber) {
     const numbers = [...winningNumber, bonusNumber];
-    if (!this.isNumbersInRange(numbers, 1, 45)) {
+    if (!isNumbersInRange(numbers, 1, 45)) {
       alert(INVALID_WINNGNUMBER_ERROR);
 
       return;
     }
-    if (!this.isDistinctNumbers(numbers)) {
+    if (!isDistinctNumbers(numbers)) {
       alert(DUPLICATED_WINNINGNUMBER_ERROR);
 
       return;
     }
-
-    this.setPrizeTable(winningNumber, bonusNumber);
-
-    const earningRate = this.calculateEarningRate();
-    this.lottoView.showPrizeTable(this.prizeTable);
-    this.lottoView.showEarningRate(earningRate);
 
     onModalShow($modal);
-  }
-
-  onSubmitManualPurchaseNumber(manaulPurcahseNumber) {
-    if (!this.isNumbersInRange(manaulPurcahseNumber, 1, 45)) {
-      alert(INVALID_WINNGNUMBER_ERROR);
-
-      return;
-    }
-    if (!this.isDistinctNumbers(manaulPurcahseNumber)) {
-      alert(DUPLICATED_WINNINGNUMBER_ERROR);
-
-      return;
-    }
-    this.lottoModel.manaulPurchase(manaulPurcahseNumber);
-    hideElement($manualPurchaseDetail);
-    this.lottoView.showPurchaseProgress(
-      this.lottoModel.price / 1000,
-      this.lottoModel.lottoList.length
-    );
-  }
-
-  onClickAutoPurchaseButton() {
-    this.lottoModel.autoPurchase(
-      this.lottoModel.price / 1000 - this.lottoModel.lottoList.length
-    );
-    this.lottoView.showConfirmation(this.lottoModel.lottoList);
-    hideElement(purchase);
-  }
-
-  onClickRestartButton() {
-    resetInput($priceInput);
-
-    this.lottoView.resetLottoView();
-    this.lottoModel.resetLottoList();
-
-    onModalClose($modal);
   }
 }
